+++ app/database/migrations/2014_02_15_012210_create_purchased_items_table.php
@@ -7,7 +7,7 @@ class CreatePurchasedItemsTable extends Migration {
     public function up() {
         Schema::create('purchased_items', function(Blueprint $table) {
             $table->integer('transaction_id')->unsigned();
-            $table->integer('barcode')->unsigned();
+            $table->bigInteger('barcode')->unsigned();
             $table->foreign('transaction_id')->references('id')->on('transactions');
             $table->foreign('barcode')->references('barcode')->on('items');
             $table->integer('quantity');
